# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAFvsOil API",
    "description": "Product scaffold for SAF vs Oil data, scenarios, and assumptions.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workspaces/{workspace_slug}/scenarios": {
      "get": {
        "tags": [
          "scenarios"
        ],
        "summary": "List Scenarios",
        "operationId": "list_scenarios_v1_workspaces__workspace_slug__scenarios_get",
        "parameters": [
          {
            "name": "workspace_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScenarioRecord"
                  },
                  "title": "Response List Scenarios V1 Workspaces  Workspace Slug  Scenarios Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "scenarios"
        ],
        "summary": "Create Scenario",
        "operationId": "create_scenario_v1_workspaces__workspace_slug__scenarios_post",
        "parameters": [
          {
            "name": "workspace_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Slug"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{workspace_slug}/scenarios/{scenario_id}": {
      "put": {
        "tags": [
          "scenarios"
        ],
        "summary": "Update Scenario",
        "operationId": "update_scenario_v1_workspaces__workspace_slug__scenarios__scenario_id__put",
        "parameters": [
          {
            "name": "workspace_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Slug"
            }
          },
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "scenarios"
        ],
        "summary": "Delete Scenario",
        "operationId": "delete_scenario_v1_workspaces__workspace_slug__scenarios__scenario_id__delete",
        "parameters": [
          {
            "name": "workspace_slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Slug"
            }
          },
          {
            "name": "scenario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Scenario Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Delete Scenario V1 Workspaces  Workspace Slug  Scenarios  Scenario Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreferencesPayload": {
        "properties": {
          "schema_version": {
            "type": "integer",
            "title": "Schema Version",
            "default": 1
          },
          "crudeSource": {
            "type": "string",
            "enum": [
              "manual",
              "brentEia",
              "brentFred"
            ],
            "title": "Crudesource",
            "default": "manual"
          },
          "carbonSource": {
            "type": "string",
            "enum": [
              "manual",
              "cbamCarbonProxyUsd"
            ],
            "title": "Carbonsource",
            "default": "manual"
          },
          "benchmarkMode": {
            "type": "string",
            "enum": [
              "crude-proxy",
              "live-jet-spot"
            ],
            "title": "Benchmarkmode",
            "default": "crude-proxy"
          },
          "crudeUsdPerBarrel": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Crudeusdperbarrel"
          },
          "carbonPriceUsdPerTonne": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Carbonpriceusdpertonne"
          },
          "subsidyUsdPerLiter": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subsidyusdperliter"
          },
          "jetProxySlope": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jetproxyslope"
          },
          "jetProxyIntercept": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jetproxyintercept"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "PreferencesPayload"
      },
      "RouteEditPayload": {
        "properties": {
          "baseCostUsdPerLiter": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Basecostusdperliter"
          },
          "co2SavingsKgPerLiter": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Co2Savingskgperliter"
          },
          "pathway": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pathway"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "RouteEditPayload"
      },
      "ScenarioCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesPayload"
          },
          "route_edits": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RouteEditPayload"
            },
            "type": "object",
            "title": "Route Edits"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ScenarioCreate"
      },
      "ScenarioRecord": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesPayload"
          },
          "route_edits": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RouteEditPayload"
            },
            "type": "object",
            "title": "Route Edits"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "workspace_slug": {
            "type": "string",
            "title": "Workspace Slug"
          },
          "saved_at": {
            "type": "string",
            "format": "date-time",
            "title": "Saved At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "workspace_slug",
          "saved_at"
        ],
        "title": "ScenarioRecord"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}